+++ src/transaction-service/Program.cs
@@ -178,9 +178,9 @@ app.UseHttpsRedirection();
 
 app.UseCorrelationId();
 
-app.UseCors();
+app.UseGlobalExceptionHandler();
 
-app.UseAuthentication();
+app.UseCors();
 app.UseAuthorization();
 
 app.MapControllers();