+++ src/transfer-service/Program.cs
@@ -145,9 +145,9 @@ app.UseHttpsRedirection();
 
 app.UseCorrelationId();
 
-app.UseCors();
+app.UseGlobalExceptionHandler();
 
-app.UseAuthentication();
+app.UseCors();
 app.UseAuthorization();
 
 app.MapControllers();