+++ src/user-service/Program.cs
@@ -124,9 +124,9 @@ app.UseHttpsRedirection();
 
 app.UseCorrelationId();
 
-app.UseCors();
+app.UseGlobalExceptionHandler();
 
-app.UseAuthentication();
+app.UseCors();
 app.UseAuthorization();
 
 app.MapControllers();